Replace the previous change with just an added method for TuberResult to enable retrieving items using a dict-like getitem method.
